VHDL-7-Segments
This project use VHDL to produce the word "BCSPN-XD" with only schematic symbols. Fot that is used karnaugh maps (Led_7_segmentos_-_BCSPN-XD.xlsx), these maps produce the logical expressions to produce each letter in the word and using the Xilinx Ide. This project was made with a Spartan E3 stater board provide by university ICESI in the subject digital logic and laboratory.